+++ frontend/src/constants.ts
@@ -8,6 +8,7 @@
 
 export * from './constantGroups/market_ui';
 export * from './constantGroups/chart_api';
+export * from './constantGroups/trader_view';
 export * from './constantGroups/features';
 export * from './constantGroups/api_auth';
 export * from './constantGroups/theme';